# gh-104078: Improve performance of PyObject_HasAttrString