@carljm 2023-05-02
How does the performance comparison look for an attribute that does exist?