@itamaro 2023-05-02
> How does the performance comparison look for an attribute that does exist?

it's about 1.2x faster

```
python -m pyperf timeit -s '
import _testcapi
hasattr_string = _testcapi.hasattr_string

class A:
  def __init__(self):
    self.attr = 1

a = A()' 'hasattr_string(a, "attr")'

# main
Mean +- std dev: 187 ns +- 4 ns

# PR
Mean +- std dev: 156 ns +- 4 ns
```