@malemburg 2023-05-07
Thanks for added insights, @carljm. I agree that it's better for `PyObject_HasAttrString()` go through `PyObject_HasAttr()`. Existence of an attribute is often easier and faster to check than actually getting the value (esp. for properties).